+++ packages/adherev-ui-prompt-errorprompt/src/index.ts
@@ -1,20 +1,24 @@
-import { message } from 'ant-design-vue';
+import { Modal, message } from 'ant-design-vue';
 
 import BfdUtil from '@baifendian/adherev-util';
 import Intl from '@baifendian/adherev-util-intl';
 
-import ErrorPrompt from './errorprompt';
+import { openErrorDialog, openErrorMessage } from './errorprompt';
 
 const {
   _util: { withVue },
 } = BfdUtil;
 
-ErrorPrompt.isUse = () => true;
-
-ErrorPrompt.use = (Vue) => {
-  Intl.isUse() && Intl.use(Vue);
-  Vue.use(message);
-  withVue(Vue, 'ErrorPrompt', ErrorPrompt);
+const ErrorPrompt = {
+  openErrorDialog,
+  openErrorMessage,
+  isUse: () => true,
+  use: (Vue) => {
+    Intl.isUse() && Intl.use(Vue);
+    Vue.use(message);
+    Vue.use(Modal);
+    withVue(Vue, 'ErrorPrompt', ErrorPrompt);
+  },
 };
 
 export default ErrorPrompt;