+++ packages/adherev-ui-prompt-successprompt/src/index.ts
@@ -1,20 +1,24 @@
-import { message } from 'ant-design-vue';
+import { Modal, message } from 'ant-design-vue';
 
 import BfdUtil from '@baifendian/adherev-util';
 import Intl from '@baifendian/adherev-util-intl';
 
-import SuccessPrompt from './successprompt';
+import { openSuccessDialog, openSuccessMessage } from './successprompt';
 
 const {
   _util: { withVue },
 } = BfdUtil;
 
-SuccessPrompt.isUse = () => true;
-
-SuccessPrompt.use = (Vue) => {
-  Intl.isUse() && Intl.use(Vue);
-  Vue.use(message);
-  withVue(Vue, 'SuccessPrompt', SuccessPrompt);
+const SuccessPrompt = {
+  openSuccessDialog,
+  openSuccessMessage,
+  isUse: () => true,
+  use: (Vue) => {
+    Intl.isUse() && Intl.use(Vue);
+    Vue.use(message);
+    Vue.use(Modal);
+    withVue(Vue, 'SuccessPrompt', SuccessPrompt);
+  },
 };
 
 export default SuccessPrompt;