+++ packages/adherev-ui-prompt-warnprompt/lib/warnprompt.d.ts
@@ -1,11 +1,14 @@
-import { MessageType } from 'ant-design-vue/types/message';
-declare const WarnPrompt: {
-    (content?: any): MessageType;
-    isUse?: () => boolean;
-    use?: (Vue: any) => void;
-};
+import { ConfigDuration, ConfigOnClose, ConfigType } from 'ant-design-vue/types/message';
+import { ModalOptions } from 'ant-design-vue/types/modal';
+/**
+ * openErrorDialog
+ * @param options
+ */
+export declare const openWarnDialog: (options: ModalOptions) => import("ant-design-vue/types/modal").ModalConfirm;
 /**
  * 警告的提示
- * @param content - {string| VNode |(h) => VNode}
+ * @param content
+ * @param duration
+ * @param onClose
  */
-export default WarnPrompt;
+export declare const openWarnMessage: (content: ConfigType, duration?: ConfigDuration | undefined, onClose?: ConfigOnClose | undefined) => import("ant-design-vue/types/message").MessageType;