+++ packages/adherev-ui-prompt-warnprompt/src/index.ts
@@ -1,20 +1,24 @@
-import { message } from 'ant-design-vue';
+import { Modal, message } from 'ant-design-vue';
 
 import BfdUtil from '@baifendian/adherev-util';
 import Intl from '@baifendian/adherev-util-intl';
 
-import WarnPrompt from './warnprompt';
+import { openWarnDialog, openWarnMessage } from './warnprompt';
 
 const {
   _util: { withVue },
 } = BfdUtil;
 
-WarnPrompt.isUse = () => true;
-
-WarnPrompt.use = (Vue) => {
-  Intl.isUse() && Intl.use(Vue);
-  Vue.use(message);
-  withVue(Vue, 'WarnPrompt', WarnPrompt);
+const WarnPrompt = {
+  openWarnMessage,
+  openWarnDialog,
+  isUse: () => true,
+  use: (Vue) => {
+    Intl.isUse() && Intl.use(Vue);
+    Vue.use(message);
+    Vue.use(Modal);
+    withVue(Vue, 'WarnPrompt', WarnPrompt);
+  },
 };
 
 export default WarnPrompt;