+++ packages/adherev-ui-prompt-warnprompt/src/warnprompt.ts
@@ -1,16 +1,29 @@
-import { message } from 'ant-design-vue';
-import { MessageType } from 'ant-design-vue/types/message';
+import { Modal, message } from 'ant-design-vue';
+import { ConfigDuration, ConfigOnClose, ConfigType } from 'ant-design-vue/types/message';
+import { ModalOptions } from 'ant-design-vue/types/modal';
 
-const WarnPrompt: {
-  (content?: any): MessageType;
-  isUse?: () => boolean;
-  use?: (Vue) => void;
-} = (content) => {
-  return message.warn(content);
-};
+import Intl from '@baifendian/adherev-util-intl';
+
+/**
+ * openErrorDialog
+ * @param options
+ */
+export const openWarnDialog = (options: ModalOptions) =>
+  Modal.warning({
+    title: Intl.tv('提示'),
+    mask: false,
+    maskClosable: true,
+    ...(options || {}),
+  });
 
 /**
  * 警告的提示
- * @param content - {string| VNode |(h) => VNode}
+ * @param content
+ * @param duration
+ * @param onClose
  */
-export default WarnPrompt;
+export const openWarnMessage = (
+  content: ConfigType,
+  duration?: ConfigDuration,
+  onClose?: ConfigOnClose,
+) => message.warning(content, duration, onClose);