+++ packages/adherev-ui-vuequill-sandbox/es/index.js
@@ -1,0 +1,2 @@
+import Util from"@baifendian/adherev-util";import VueQuillSandbox from"./vueQuillSandbox";var _a=Util._util,withInstall=_a.withInstall,withVue=_a.withVue,Component=withInstall(VueQuillSandbox);Component.isUse=function(){return!0},Component.use=function(n){n.use(Component),withVue(n,"VueQuillSandbox",Component)};export default Component;
+//# sourceMappingURL=index.js.map