+++ packages/adherev-ui-vuequill-sandbox/es/vueQuillSandbox.d.ts
@@ -1,0 +1,2 @@
+declare const VueQuillSandbox: any;
+export default VueQuillSandbox;