+++ packages/adherev-ui-vuequill-sandbox/src/demo/index.tsx
@@ -1,0 +1,16 @@
+// import Vue from 'vue';
+//
+// import VueQuillSandbox from '../index';
+// // @ts-ignore
+// import Quill from './quill';
+//
+// import '../index.less';
+//
+// VueQuillSandbox.use(Vue);
+//
+// new Vue({
+//   el: '#app',
+//   render(h) {
+//     return h(Quill);
+//   },
+// });