+++ packages/adherev-ui-vuequill-sandbox/src/index.ts
@@ -1,0 +1,18 @@
+import Util from '@baifendian/adherev-util';
+
+import VueQuillSandbox from './vueQuillSandbox';
+
+const {
+  _util: { withInstall, withVue },
+} = Util;
+
+const Component = withInstall(VueQuillSandbox);
+
+Component.isUse = () => true;
+
+Component.use = (Vue) => {
+  Vue.use(Component);
+  withVue(Vue, 'VueQuillSandbox', Component);
+};
+
+export default Component;